separated the cli into host and vm modules. Still need to be moved to subcommand
