+++ src/cli.rs
@@ -11,14 +11,13 @@
 //! and executes the corresponding subcommand.
 
 use std::path::PathBuf;
-use tokio::fs::File;
 
-use anyhow::{Context, Result};
+use anyhow::{Result};
 use clap_derive::Parser;
 
 use crate::{
     logging::LoggingFormat,
-    tricks::{report::Report, status::Status, Trick},
+    host,
 };
 
 /// Describes Houdini's command line interface.
@@ -46,6 +45,8 @@ enum Cmd {
         #[clap(min_values = 1, required = true)]
         exploits: Vec<PathBuf>,
     },
+
+    Vm {},
 }
 
 impl Cli {
@@ -53,36 +54,11 @@ impl Cli {
     pub async fn run(self) -> Result<()> {
         match self.subcmd {
             Cmd::Run { exploits } => {
-                let mut report = Report::default();
-
-                for exploit in exploits {
-                    let f = File::open(&exploit).await.context(format!(
-                        "could not open exploit file {}",
-                        &exploit.display()
-                    ))?;
-                    let plan: Trick = serde_yaml::from_reader(f.into_std().await).context(
-                        format!("failed to parse exploit file {}", &exploit.display()),
-                    )?;
-
-                    let status = plan.run(Some(&mut report)).await;
-                    match status {
-                        Status::Undecided | Status::SetupFailure | Status::ExploitFailure => {
-                            tracing::info!(status = ?status, "plan execution FAILED");
-                        }
-                        Status::ExploitSuccess => {
-                            tracing::info!(status = ?status, "plan execution SUCCEEDED");
-                        }
-                        Status::Skip => {
-                            tracing::info!(status = ?status, "plan execution SKIPPED");
-                        }
-                    }
-                }
-
-                report
-                    .write_to_disk()
-                    .await
-                    .context("failed to write report to disk")?;
+                host::main(exploits).await?;
             }
+            Cmd::Vm { } => {
+                
+            } 
         }
 
         Ok(())
