+++ src/subcommands.rs
@@ -1,0 +1,5 @@
+/// TODO: Add host and vm subcommands here?
+/// 
+fn main(){
+    
+}
