+++ src/vm.rs
@@ -1,0 +1,8 @@
+/// TODO: logic for running vm code goes here
+fn main(){
+    // launch and start a unix socket 
+    // wait for step from host
+    // run steps recieved from host
+    // send back step output
+    // exit when receiving some form of exploit finished signal
+}
